task to tsx and adding usecallback
